+++ Model/Config.php
@@ -26,6 +26,9 @@ class Config implements ConfigInterface, DebugConfigInterface
     public const XML_PATH_CSP_DEBUG_MODE_ENABLED = 'csp/general/debug_mode_enabled';
     public const XML_PATH_CSP_VALUE_OPTIMIZATION_ENABLED = 'csp/general/enable_value_optimization';
     public const XML_PATH_CSP_REDUNDANT_WILDCARD_REMOVAL_ENABLED = 'csp/general/enable_redundant_wildcard_removal';
+    public const XML_PATH_CSP_REPORT_CLEANUP_ENABLED = 'csp/report_cleanup/enabled';
+    public const XML_PATH_CSP_REPORT_CLEANUP_MODE = 'csp/report_cleanup/mode';
+    public const XML_PATH_CSP_REPORT_CLEANUP_THRESHOLD = 'csp/report_cleanup/threshold';
 
 
     public function __construct(private readonly ScopeConfigInterface $scopeConfig)
@@ -126,6 +129,30 @@ class Config implements ConfigInterface, DebugConfigInterface
     public function isRedundantWildcardRemovalEnabled(): bool
     {
         return $this->scopeConfig->isSetFlag(self::XML_PATH_CSP_REDUNDANT_WILDCARD_REMOVAL_ENABLED);
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function isReportCleanupEnabled(): bool
+    {
+        return $this->scopeConfig->isSetFlag(self::XML_PATH_CSP_REPORT_CLEANUP_ENABLED);
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function getReportCleanupMode(): string
+    {
+        return (string)($this->scopeConfig->getValue(self::XML_PATH_CSP_REPORT_CLEANUP_MODE) ?: 'date');
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function getReportCleanupThreshold(): int
+    {
+        return (int)($this->scopeConfig->getValue(self::XML_PATH_CSP_REPORT_CLEANUP_THRESHOLD) ?: 30);
     }
 }